Do not always unroll the permutation loop
This makes the code above 100ko for a degree of 16k. Moreover, some
preliminary tests show that there is a performance benefits only above
2048. It seems to be the limit of the L1 instruction cache on x86
(16Ko). The limit is set here to 1024 not to always fill the instruction
cache with the permutation code, and thus have a potential penalty with
the rest of the code. A vectorized version that do not use a precomputed
permutation also shows great potential (not in this commit).

This also has the benefits to speed up the compilation process for
degrees above 1024.